+++ CHANGELOG.md
@@ -1,6 +1,7 @@
  == CHANGELOG ==
 
 * The screenshot dialog now saves the image format and path selected as settings.
+* Added limited support for rendering Starfield flipbooks and layered emissivity, and for the use vertex color as tint material setting.
 * Fixed bugs in rendering ordered nodes.
 * Fixed the axes not being correctly drawn depending on the OpenGL settings from the last shape.
 * Fixes in the lighting only and textures disabled rendering modes.